+++ src/enums/headers.enum.ts
@@ -1,0 +1,9 @@
+export enum Headers {
+    nameLabel = 'nameLabel',
+    genderLabel = 'genderLabel',
+    citizenshipLabel = 'citizenshipLabel',
+    skillsLabel = 'skillsLabel',
+    occupationLabel = 'occupationLabel',
+    memberOfLabel = 'memberOfLabel',
+    creatorLabel = 'creatorLabel'
+}